ec2_deploy/requirements.txt
pandas>=2.0
google-cloud-bigquery>=3.0
google-auth>=2.0
python-dotenv>=1.0
requests>=2.31
redshift-connector>=2.0
psycopg2-binary>=2.9
paramiko==3.5.0
sshtunnel>=0.4
